Replace old store file
