+++ webui/react/src/components/DeterminedAuth.tsx
@@ -16,7 +16,7 @@ import { ErrorType } from 'shared/utils/error';
 import { StorageManager } from 'shared/utils/storage';
 import { setAuth } from 'stores/auth';
 import { PermissionsStore } from 'stores/permissions';
-import usersStore from 'stores/usersObserve';
+import usersStore from 'stores/users';
 import handleError from 'utils/error';
 
 import css from './DeterminedAuth.module.scss';
