+++ webui/react/src/components/NavigationTabbar.tsx
@@ -15,7 +15,7 @@ import { AnyMouseEvent, routeToReactUrl } from 'shared/utils/routes';
 import { selectIsAuthenticated } from 'stores/auth';
 import { useClusterStore } from 'stores/cluster';
 import { initInfo, useDeterminedInfo } from 'stores/determinedInfo';
-import usersStore from 'stores/usersObserve';
+import usersStore from 'stores/users';
 import { useWorkspaces } from 'stores/workspaces';
 import { BrandingType } from 'types';
 import { Loadable } from 'utils/loadable';
