+++ webui/react/src/hooks/useModal/UserSettings/useModalPasswordChange.test.tsx
@@ -7,7 +7,7 @@ import { V1LoginRequest } from 'services/api-ts-sdk';
 import { SetUserPasswordParams } from 'services/types';
 import { StoreProvider as UIProvider } from 'shared/contexts/stores/UI';
 import { setAuth } from 'stores/auth';
-import usersStore from 'stores/usersObserve';
+import usersStore from 'stores/users';
 import { DetailedUser } from 'types';
 
 import useModalPasswordChange, {
