+++ webui/react/src/pages/ModelRegistry.tsx
@@ -50,7 +50,7 @@ import { isEqual } from 'shared/utils/data';
 import { ErrorType } from 'shared/utils/error';
 import { validateDetApiEnum } from 'shared/utils/service';
 import { alphaNumericSorter } from 'shared/utils/sort';
-import usersStore from 'stores/usersObserve';
+import usersStore from 'stores/users';
 import { useEnsureWorkspacesFetched, useWorkspaces } from 'stores/workspaces';
 import { DetailedUser, ModelItem, Workspace } from 'types';
 import handleError from 'utils/error';
