+++ webui/react/src/pages/Settings/SettingsAccount.test.tsx
@@ -7,7 +7,7 @@ import { NEW_PASSWORD_LABEL } from 'hooks/useModal/UserSettings/useModalPassword
 import { PatchUserParams } from 'services/types';
 import { StoreProvider as UIProvider } from 'shared/contexts/stores/UI';
 import { setAuth } from 'stores/auth';
-import usersStore from 'stores/usersObserve';
+import usersStore from 'stores/users';
 import { DetailedUser } from 'types';
 
 import SettingsAccount, { CHANGE_PASSWORD_TEXT } from './SettingsAccount';
