+++ webui/react/src/pages/TrialsComparison/Collections/useTrialCollections.tsx
@@ -13,7 +13,7 @@ import { deleteTrialsCollection, getTrialsCollections, patchTrialsCollection } f
 import Icon from 'shared/components/Icon';
 import { clone, finiteElseUndefined, isFiniteNumber } from 'shared/utils/data';
 import { ErrorType } from 'shared/utils/error';
-import usersStore from 'stores/usersObserve';
+import usersStore from 'stores/users';
 import handleError from 'utils/error';
 import { Loadable } from 'utils/loadable';
