+++ webui/react/src/pages/TrialsComparison/Table/TrialTable.tsx
@@ -26,7 +26,7 @@ import { paths } from 'routes/utils';
 import { Trialv1State, V1AugmentedTrial } from 'services/api-ts-sdk';
 import { ColorScale, glasbeyColor } from 'shared/utils/color';
 import { isFiniteNumber } from 'shared/utils/data';
-import usersStore from 'stores/usersObserve';
+import usersStore from 'stores/users';
 import { StateOfUnion } from 'themes';
 import { DetailedUser, MetricType } from 'types';
 import { Loadable } from 'utils/loadable';
