+++ webui/react/src/stores/users.tsx
@@ -1,9 +1,10 @@
 import { Map } from 'immutable';
-import React, { createContext, PropsWithChildren, useCallback, useContext, useState } from 'react';
+import { Observable, observable, WritableObservable } from 'micro-observables';
 
 import { getCurrentUser, getUsers } from 'services/api';
-import { V1GetUsersRequestSortBy, V1Pagination } from 'services/api-ts-sdk';
-import { DetailedUser } from 'types';
+import { V1Pagination } from 'services/api-ts-sdk';
+import type { GetUsersParams as FetchUsersConfig } from 'services/types';
+import { DetailedUser, DetailedUserList } from 'types';
 import handleError from 'utils/error';
 import { Loadable, Loaded, NotLoaded } from 'utils/loadable';
 import { encodeParams } from 'utils/store';
@@ -13,230 +14,108 @@ type UsersPagination = {
   users: number[];
 };
 
-export type UserPage = {
-  pagination: V1Pagination;
-  users: DetailedUser[];
-};
+class UsersService {
+  #users: WritableObservable<Map<number, DetailedUser>> = observable(Map());
+  #usersByKey: WritableObservable<Map<string, UsersPagination>> = observable(Map());
+  #currentUserId: WritableObservable<Loadable<number>> = observable(NotLoaded);
 
-type UsersContext = {
-  currentUser: Loadable<number>;
-  updateCurrentUser: (fn: (currentUser: Loadable<number>) => Loadable<number>) => void;
-  updateUsers: (fn: (users: Map<number, DetailedUser>) => Map<number, DetailedUser>) => void;
-  updateUsersByKey: (
-    fn: (users: Map<string, UsersPagination>) => Map<string, UsersPagination>,
-  ) => void;
-  users: Map<number, DetailedUser>;
-  usersByKey: Map<string, UsersPagination>;
-};
+  public getUser = (id: number): Observable<Loadable<DetailedUser>> => {
+    return this.#users.select((map) => {
+      const user = map.get(id);
+      return user ? Loaded(user) : NotLoaded;
+    });
+  };
 
-export type FetchUsersConfig = {
-  limit: number;
-  offset: number;
-  orderBy: 'ORDER_BY_DESC' | 'ORDER_BY_ASC';
-  sortBy: V1GetUsersRequestSortBy;
-};
+  public getCurrentUser = (): Observable<Loadable<DetailedUser>> => {
+    return this.#users.select((map) => {
+      const id = this.#currentUserId.get();
+      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+      return Loadable.map(id, (id) => map.get(id)!);
+    });
+  };
 
-const UsersContext = createContext<UsersContext | null>(null);
-
-export const UsersProvider: React.FC<PropsWithChildren> = ({ children }) => {
-  const [usersByKey, setUsersByKey] = useState<Map<string, UsersPagination>>(
-    Map<string, UsersPagination>(),
-  );
-  const [users, setUsers] = useState<Map<number, DetailedUser>>(Map<number, DetailedUser>());
-  const [currentUser, setCurrentUser] = useState<Loadable<number>>(NotLoaded);
-
-  return (
-    <UsersContext.Provider
-      value={{
-        currentUser,
-        updateCurrentUser: setCurrentUser,
-        updateUsers: setUsers,
-        updateUsersByKey: setUsersByKey,
-        users,
-        usersByKey,
-      }}>
-      {children}
-    </UsersContext.Provider>
-  );
-};
-
-export const useFetchUsers = (canceler: AbortController): ((cfg?: FetchUsersConfig) => void) => {
-  const context = useContext(UsersContext);
-
-  if (context === null) {
-    throw new Error('Attempted to use useFetchUsers outside of Users Context');
-  }
-
-  const { updateUsersByKey, updateUsers } = context;
-
-  return useCallback(
-    async (cfg?: FetchUsersConfig) => {
-      try {
-        const config = cfg ?? {};
-        const response = await getUsers(config, { signal: canceler.signal });
-        const usersPages = {
-          pagination: response.pagination,
-          users: response.users.map((user) => user.id),
-        };
-
-        updateUsersByKey((prevState) => prevState.set(encodeParams(config), usersPages));
-        updateUsers((prevState) => {
-          return prevState.withMutations((map) => {
-            response.users.forEach((user) => map.set(user.id, user));
-          });
-        });
-      } catch (e) {
-        handleError(e, { publicSubject: 'Unable to fetch users.' });
-      }
-    },
-    [canceler, updateUsers, updateUsersByKey],
-  );
-};
-
-export const useEnsureCurrentUserFetched = (canceler: AbortController): (() => Promise<void>) => {
-  const context = useContext(UsersContext);
-
-  if (context === null) {
-    throw new Error('Attempted to use useEnsureCurrentUserFetched outside of Users Context');
-  }
-
-  const { updateCurrentUser, currentUser, updateUsers } = context;
-
-  return useCallback(async (): Promise<void> => {
-    if (currentUser !== NotLoaded) return;
+  public ensureCurrentUserFetched = async (canceler: AbortController): Promise<void> => {
+    if (this.#currentUserId.get() !== NotLoaded) return;
 
     try {
       const response = await getCurrentUser({ signal: canceler.signal });
 
-      updateUsers((prevState) => prevState.set(response.id, response));
-      updateCurrentUser(() => Loaded(response.id));
+      this.updateUsers([response]);
+      this.#currentUserId.set(Loaded(response.id));
     } catch (e) {
-      handleError(e);
+      handleError(e, { publicSubject: 'Unable to fetch current user.' });
     }
-  }, [canceler, updateCurrentUser, currentUser, updateUsers]);
-};
-
-export const useEnsureUsersFetched = (canceler: AbortController): (() => Promise<void>) => {
-  const context = useContext(UsersContext);
-
-  if (context === null) {
-    throw new Error('Attempted to use useFetchUsers outside of Users Context');
-  }
-
-  const { updateUsers, updateUsersByKey, usersByKey } = context;
-
-  return useCallback(
-    async (cfg?: FetchUsersConfig): Promise<void> => {
-      const config = cfg ?? {};
-      const usersPagination = usersByKey.get(encodeParams(config));
-
-      if (usersPagination) return;
-
-      try {
-        const response = await getUsers(config, { signal: canceler.signal });
-        const usersPages = {
-          pagination: response.pagination,
-          users: response.users.map((user) => user.id),
-        };
-
-        updateUsersByKey((prevState) => prevState.set(encodeParams(config), usersPages));
-        updateUsers((prevState) => {
-          return prevState.withMutations((map) => {
-            response.users.forEach((user) => map.set(user.id, user));
-          });
-        });
-      } catch (e) {
-        handleError(e);
-      }
-    },
-    [canceler, updateUsers, usersByKey, updateUsersByKey],
-  );
-};
-
-export const useUsers = (cfg?: FetchUsersConfig): Loadable<Readonly<UserPage>> => {
-  const context = useContext(UsersContext);
-
-  if (context === null) {
-    throw new Error('Attempted to use useUsers outside of Users Context');
-  }
-  const config = cfg ?? {};
-  const usersPagination = context.usersByKey.get(encodeParams(config));
-
-  if (!usersPagination) return NotLoaded;
-
-  const userPage: UserPage = {
-    pagination: usersPagination.pagination,
-    users: usersPagination.users.flatMap((userId) => {
-      const user = context.users.get(userId);
-
-      return user ? [user] : [];
-    }),
   };
 
-  return Loaded(userPage);
-};
+  public updateCurrentUser = (id: number | null) => {
+    if (id === null) this.#currentUserId.set(NotLoaded);
+    else this.#currentUserId.set(Loaded(id));
+  };
 
-export const useUpdateCurrentUser = (): ((id: number | null) => void) => {
-  const context = useContext(UsersContext);
+  public getUsers = (cfg?: FetchUsersConfig): Observable<Loadable<DetailedUserList>> => {
+    const config = cfg ?? {};
 
-  if (context === null) {
-    throw new Error('Attempted to use useUpdateCurrentUser outside of Users Context');
-  }
+    return this.#usersByKey.select((map) => {
+      const usersPagination = map.get(encodeParams(config));
 
-  const { updateCurrentUser } = context;
-  const callback = useCallback(
-    (id: number | null) => {
-      if (id === null) {
-        updateCurrentUser(() => NotLoaded);
-      } else {
-        updateCurrentUser(() => Loaded(id));
-      }
-    },
-    [updateCurrentUser],
-  );
+      if (!usersPagination) return NotLoaded;
 
-  return callback;
-};
+      const userPage: DetailedUserList = {
+        pagination: usersPagination.pagination,
+        users: usersPagination.users.flatMap((userId) => {
+          const user = this.#users.get().get(userId);
 
-export const useUpdateUser = (): ((
-  id: number,
-  updater: (arg0: DetailedUser) => DetailedUser,
-) => void) => {
-  const context = useContext(UsersContext);
+          return user ? [user] : [];
+        }),
+      };
 
-  if (context === null) {
-    throw new Error('Attempted to use useUpdateUser outside of Users Context');
-  }
+      return Loaded(userPage);
+    });
+  };
 
-  const { updateUsers } = context;
-  const callback = useCallback(
-    (id: number, updater: (arg0: DetailedUser) => DetailedUser) => {
-      updateUsers((prevState) => {
-        if (prevState.has(id)) {
-          // this state is statically guaranteed to be non-null
-          // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-          return prevState.update(id, (detailedUser) => updater(detailedUser!));
-        }
+  public ensureUsersFetched = async (
+    canceler: AbortController,
+    cfg?: FetchUsersConfig,
+  ): Promise<void> => {
+    const config = cfg ?? {};
+    const usersPagination = this.#usersByKey.get().get(encodeParams(config));
 
-        return prevState;
+    if (usersPagination) return;
+
+    try {
+      const response = await getUsers(config, { signal: canceler?.signal });
+
+      this.updateUsersByKey(config, response);
+      this.updateUsers(response.users);
+    } catch (e) {
+      handleError(e, { publicSubject: 'Unable to fetch users.' });
+    }
+  };
+
+  public updateUsers = (users: DetailedUser | DetailedUser[]) => {
+    this.#users.update((map) => {
+      return map.withMutations((map) => {
+        if (Array.isArray(users)) users.forEach((user) => map.set(user.id, user));
+        else map.set(users.id, users);
       });
-    },
-    [updateUsers],
-  );
+    });
+  };
 
-  return callback;
-};
+  private updateUsersByKey = (
+    config: FetchUsersConfig | Record<string, never>,
+    usersList: DetailedUserList,
+  ) => {
+    const usersPages = {
+      pagination: usersList.pagination,
+      users: usersList.users.map((user) => user.id),
+    };
 
-export const useCurrentUser = (): Loadable<DetailedUser> => {
-  const context = useContext(UsersContext);
+    this.#usersByKey.update((map) => map.set(encodeParams(config), usersPages));
+  };
+}
 
-  if (context === null) {
-    throw new Error('Attempted to use useCurrentUser outside of User Context');
-  }
-  const { currentUser, users } = context; // this state is statically guaranteed to be non-null
+const usersStore = new UsersService();
 
-  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-  const loadedUser = Loadable.map(currentUser, (userId) => users.get(userId)!);
+export { FetchUsersConfig };
 
-  return loadedUser;
-};
+export default usersStore;
